♻️ : user, term, auth 관련 dto 파일 분리 및 파일 이름 수정
