+++ .gitignore
@@ -57,3 +57,5 @@ report.[0-9]*.[0-9]*.[0-9]*.[0-9]*.json
 
 # local DB caches
 /db_data
+
+.cursorrules
