+++ src/modules/auth/dto/internal/local-auth/check-local-auth.dto.ts
@@ -1,6 +1,6 @@
 import { Expose } from 'class-transformer';
 import { IsNumber } from 'class-validator';
-import { BaseInternalDto } from 'src/shared/dtos/base-internal-dto';
+import { BaseInternalDto } from 'src/shared/dtos/base-internal.dto';
 
 export class CheckLocalAuthDto extends BaseInternalDto {
   @Expose({ name: 'user_id' })
