+++ src/modules/auth/dto/internal/local-auth/update-local-auth.dto.ts
@@ -1,5 +1,5 @@
 import { IsNotEmpty, IsNumber, IsString } from 'class-validator';
-import { BaseInternalDto } from 'src/shared/dtos/base-internal-dto';
+import { BaseInternalDto } from 'src/shared/dtos/base-internal.dto';
 
 export class UpdateLocalAuthDto extends BaseInternalDto {
   @IsNotEmpty()
