+++ src/modules/auth/dto/internal/social-auth/findone-social-auth.dto.ts
@@ -1,6 +1,6 @@
 import { IsEnum, IsNotEmpty, IsNumber, IsString } from 'class-validator';
 import { SocialProvider } from 'src/modules/auth/const/auth.const';
-import { BaseInternalDto } from 'src/shared/dtos/base-internal-dto';
+import { BaseInternalDto } from 'src/shared/dtos/base-internal.dto';
 import { Transform } from 'class-transformer';
 
 export class FindOneResultSocialAuthDto extends BaseInternalDto {
