+++ src/modules/team/dto/internal/team-select-id.dto.ts
@@ -1,5 +1,5 @@
 import { IsNotEmpty, IsNumber } from 'class-validator';
-import { BaseInternalDto } from 'src/shared/dtos/base-internal-dto';
+import { BaseInternalDto } from 'src/shared/dtos/base-internal.dto';
 
 export class TeamSelectIdDto extends BaseInternalDto {
   @IsNumber()
