+++ src/modules/user/dto/internal/create-user.dto.ts
@@ -1,4 +1,4 @@
-import { BaseInternalDto } from 'src/shared/dtos/base-internal-dto';
+import { BaseInternalDto } from 'src/shared/dtos/base-internal.dto';
 import {
   IsEmail,
   IsNotEmpty,
