+++ src/modules/user/dto/internal/user-with-local-auth.dto.ts
@@ -7,10 +7,10 @@ import {
   IsString,
   ValidateNested,
 } from 'class-validator';
-import { BaseInternalDto } from 'src/shared/dtos/base-internal-dto';
+import { BaseInternalDto } from 'src/shared/dtos/base-internal.dto';
 import { CheckLocalAuthDto } from 'src/modules/auth/dto/internal/local-auth/check-local-auth.dto';
 
-export class FindOneWithLocalAuthDto extends BaseInternalDto {
+export class UserWithLocalAuthDto extends BaseInternalDto {
   @IsNotEmpty()
   @IsNumber()
   id: number;
