+++ src/modules/user/dto/internal/user-with-team.dto.ts
@@ -1,9 +1,9 @@
 import { Type } from 'class-transformer';
 import { IsEmail, IsNotEmpty, IsNumber, IsString } from 'class-validator';
-import { BaseInternalDto } from 'src/shared/dtos/base-internal-dto';
-import { TeamDto } from 'src/modules/team/dto/response/team.dto';
+import { BaseInternalDto } from 'src/shared/dtos/base-internal.dto';
+import { TeamDto } from 'src/modules/team/dto/response/res-team.dto';
 
-export class FindOneUserWithTeamDto extends BaseInternalDto {
+export class UserWithTeamDto extends BaseInternalDto {
   @IsNotEmpty()
   @IsNumber()
   id: number;
