+++ src/modules/user/dto/request/req-create-local-user.dto.ts
@@ -1,0 +1,32 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { IsEmail, IsNotEmpty, IsNumber, IsString } from 'class-validator';
+
+export class CreateLocalUserDto {
+  @ApiProperty({
+    example: 'test@test.com',
+  })
+  @IsNotEmpty()
+  @IsString()
+  @IsEmail()
+  email: string;
+
+  @ApiProperty({
+    example: 'should be hidden',
+  })
+  @IsNotEmpty()
+  @IsString()
+  password: string;
+
+  @IsString()
+  image: string;
+
+  @IsString()
+  nickname: string;
+
+  @ApiProperty({
+    example: 1,
+  })
+  @IsNotEmpty()
+  @IsNumber()
+  teamId: number;
+}
