+++ src/modules/user/dto/request/req-nickname-user-dto.ts
@@ -1,0 +1,9 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { IsNotEmpty, IsString } from 'class-validator';
+
+export class NicknameDto {
+  @ApiProperty({})
+  @IsNotEmpty()
+  @IsString()
+  nickname: string;
+}
