+++ src/shared/dtos/base-internal.dto.ts
